Centralize schema/type in component
