+++ apps/prairielearn/src/components/AssessmentUsageModal.tsx
@@ -1,7 +1,24 @@
 import { useMemo } from 'react';
 import { Modal } from 'react-bootstrap';
+import { z } from 'zod';
 
-import type { AssessmentUsage } from '../lib/client/assessment-usage.js';
+import { IdSchema } from '@prairielearn/zod';
+
+/**
+ * An assessment as shown in "View assessments" usage lists, with enough course
+ * instance context to group it and link to it.
+ */
+export const AssessmentUsageSchema = z.object({
+  assessment_id: IdSchema,
+  tid: z.string(),
+  title: z.string(),
+  label: z.string(),
+  color: z.string(),
+  course_instance_id: IdSchema,
+  course_instance_short_name: z.string().nullable(),
+  course_instance_long_name: z.string().nullable(),
+});
+export type AssessmentUsage = z.infer<typeof AssessmentUsageSchema>;
 
 /**
  * Lists the assessments that reference a course entity (e.g. an assessment set
